Reorganized code to be modular
