+++ Image-Path/Image/ECE-4440-Color-Detection.py
